WIP: added interface for fixed texture rendering.

+++ src/Engine/Graphics/Textures/texture.hpp
@@ -25,6 +25,6 @@ public:
 private:
     friend class TextureManager;
     Texture (size_t idIn, TextureManager* managerIn);
-    unsigned int id;
+    size_t id;
     TextureManager* manager;
 };
